+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * custom arrowWidget on configuration.
 * add custom gap on json list.
 * `openAtStart` option. (original post #5)
-* fix null value text style
+* fix null value text style.
+* now can expand stringTile. #10
 
 ## 0.3.1